refactor: moved methods to documentloader file and removed applyCloudSQLDocumentLoaderOptions method
